kishi rebuild: clean drivetrain + shooter + intake only
