+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Intake.java
@@ -1,53 +1,77 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.kishi;
 
+import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
+import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
-// intake subsystem — 3 states: off, forward, reverse
+/**
+ * Intake.java — Ball Intake Subsystem
+ *
+ * HOW IT WORKS:
+ * -------------
+ * One motor spins a roller or set of wheels that pulls balls
+ * into the robot from the field. Running it in reverse ejects balls.
+ *
+ * HARDWARE INDEPENDENCE:
+ * ----------------------
+ * Motor reference comes in via HardwareMapConfig — this class
+ * never calls HardwareMap.get() directly.
+ */
 public class Intake {
 
-    public enum IntakeState {
-        OFF,
-        INTAKE,
-        REVERSE
-    }
-
-    private static final double INTAKE_POWER  =  1.0;
-    private static final double REVERSE_POWER = -1.0;
+    // Motor power for intake and eject
+    private static final double INTAKE_POWER = 0.85;
+    private static final double EJECT_POWER  = -0.6;
 
     private final DcMotorEx motor;
-    private IntakeState state;
 
+    // ----------------------------------
+    // CONSTRUCTOR
+    // ----------------------------------
+
+    /**
+     * @param robot Shared HardwareMapConfig with intake_motor wired up
+     */
     public Intake(HardwareMapConfig robot) {
         this.motor = robot.intake_motor;
-        this.motor.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
-        this.motor.setZeroPowerBehavior(DcMotorEx.ZeroPowerBehavior.BRAKE);
-        this.state = IntakeState.OFF;
+
+        // Direction: FORWARD = pull balls in. Flip to REVERSE if intake runs backwards.
+        motor.setDirection(DcMotorSimple.Direction.FORWARD);
+
+        // Raw power control — no encoder needed for intake
+        motor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+
+        // BRAKE so the roller stops instantly when power is cut,
+        // preventing balls from rolling back out
+        motor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
     }
 
-    public void intakeIn() {
-        state = IntakeState.INTAKE;
+    // ----------------------------------
+    // CONTROL METHODS
+    // ----------------------------------
+
+    /** Pull balls into the robot */
+    public void intake() {
         motor.setPower(INTAKE_POWER);
     }
 
-    public void intakeReverse() {
-        state = IntakeState.REVERSE;
-        motor.setPower(REVERSE_POWER);
+    /** Push balls back out (jam clearing or overflow) */
+    public void eject() {
+        motor.setPower(EJECT_POWER);
     }
 
-    public void intakeOff() {
-        state = IntakeState.OFF;
+    /** Stop the intake roller */
+    public void stop() {
         motor.setPower(0);
     }
 
-    public void stop() {
-        intakeOff();
+    /** Returns true if the intake is currently running */
+    public boolean isRunning() {
+        return Math.abs(motor.getPower()) > 0.01;
     }
 
-    public IntakeState getState() {
-        return state;
-    }
-
-    public double getVelocity() {
-        return motor.getVelocity();
+    /** Displays current intake power to Driver Station */
+    public void displayTelemetry(org.firstinspires.ftc.robotcore.external.Telemetry telemetry) {
+        telemetry.addData("Intake | power", String.format("%.2f", motor.getPower()));
     }
 }
