--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/webcam_test/src/AprilTagProcessor.java
@@ -1,346 +1,0 @@
-/**
- * AprilTagProcessor.java
- * 
- * Performs AprilTag detection on camera frames.
- * Handles tag detection, pose estimation, and data extraction.
- * 
- * This class wraps the FTC Vision AprilTag detection library and provides
- * a simplified interface for detecting tags in video frames.
- */
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-
-/**
- * Processes video frames to detect AprilTags
- */
-public class AprilTagProcessor {
-    
-    // ============================================================================
-    // CONFIGURATION CONSTANTS
-    // ============================================================================
-    
-    /** AprilTag decimation factor (1=accurate, 4=fast) */
-    private int decimation;
-    
-    /** AprilTag family to detect */
-    private static final String TAG_FAMILY = "TAG_36h11";
-    
-    /** Detection confidence threshold (0.0 to 1.0) */
-    private static final double DETECTION_THRESHOLD = 0.5;
-    
-    /** Output units for position (INCHES, CENTIMETERS, METERS) */
-    private static final String OUTPUT_UNITS = "INCHES";
-    
-    /** Use fast pose estimation */
-    private static final boolean USE_FAST_POSE = false;
-    
-    /** Enable debug output */
-    private static final boolean DEBUG = true;
-    
-    // ============================================================================
-    // INSTANCE VARIABLES
-    // ============================================================================
-    
-    /** AprilTag detector object (would be actual detector in real implementation) */
-    private Object detector;
-    
-    /** Flag indicating if processor is initialized */
-    private boolean isInitialized;
-    
-    /** Statistics */
-    private long framesProcessed;
-    private long tagsDetected;
-    private long lastProcessTime;
-    
-    /** Random number generator for simulation */
-    private Random random;
-    
-    // ============================================================================
-    // CONSTRUCTOR
-    // ============================================================================
-    
-    /**
-     * Create a new AprilTagProcessor
-     * 
-     * @param decimation Decimation factor (1-4, lower = more accurate)
-     */
-    public AprilTagProcessor(int decimation) {
-        this.decimation = Math.max(1, Math.min(4, decimation));
-        this.isInitialized = false;
-        this.framesProcessed = 0;
-        this.tagsDetected = 0;
-        this.lastProcessTime = System.currentTimeMillis();
-        this.random = new Random();
-        
-        if (DEBUG) {
-            System.out.println("[AprilTagProcessor] Created with decimation " + this.decimation);
-        }
-    }
-    
-    // ============================================================================
-    // INITIALIZATION METHODS
-    // ============================================================================
-    
-    /**
-     * Initialize the AprilTag processor
-     * 
-     * @return true if initialization successful, false otherwise
-     */
-    public boolean initialize() {
-        try {
-            if (DEBUG) {
-                System.out.println("[AprilTagProcessor] Initializing AprilTag processor...");
-            }
-            
-            // In a real implementation, this would:
-            // 1. Create AprilTagProcessor instance
-            // 2. Set tag family
-            // 3. Set decimation
-            // 4. Configure pose estimation
-            // 5. Set output units
-            
-            if (!initializeDetector()) {
-                System.err.println("[AprilTagProcessor] Failed to initialize detector");
-                return false;
-            }
-            
-            isInitialized = true;
-            
-            if (DEBUG) {
-                System.out.println("[AprilTagProcessor] Processor initialized successfully");
-                System.out.println("  - Tag Family: " + TAG_FAMILY);
-                System.out.println("  - Decimation: " + decimation);
-                System.out.println("  - Output Units: " + OUTPUT_UNITS);
-            }
-            
-            return true;
-            
-        } catch (Exception e) {
-            System.err.println("[AprilTagProcessor] Initialization error: " + e.getMessage());
-            e.printStackTrace();
-            return false;
-        }
-    }
-    
-    /**
-     * Initialize the detector
-     * 
-     * @return true if successful
-     */
-    private boolean initializeDetector() {
-        try {
-            // In a real implementation, this would create the actual AprilTag detector
-            // For now, we simulate it
-            detector = new Object(); // Placeholder
-            return true;
-        } catch (Exception e) {
-            System.err.println("[AprilTagProcessor] Detector initialization failed: " + e.getMessage());
-            return false;
-        }
-    }
-    
-    // ============================================================================
-    // FRAME PROCESSING METHODS
-    // ============================================================================
-    
-    /**
-     * Process a frame to detect AprilTags
-     * 
-     * @param frameData Frame data as byte array
-     * @return List of detected AprilTags
-     */
-    public List<AprilTagDetection> processFrame(byte[] frameData) {
-        if (!isInitialized || frameData == null) {
-            return new ArrayList<>();
-        }
-        
-        try {
-            long startTime = System.currentTimeMillis();
-            
-            // In a real implementation, this would:
-            // 1. Convert frame data to appropriate format
-            // 2. Run AprilTag detection algorithm
-            // 3. Extract detection results
-            // 4. Calculate pose estimation
-            // 5. Return list of detections
-            
-            List<AprilTagDetection> detections = detectTags(frameData);
-            
-            framesProcessed++;
-            tagsDetected += detections.size();
-            lastProcessTime = System.currentTimeMillis() - startTime;
-            
-            return detections;
-            
-        } catch (Exception e) {
-            System.err.println("[AprilTagProcessor] Frame processing error: " + e.getMessage());
-            return new ArrayList<>();
-        }
-    }
-    
-    /**
-     * Detect AprilTags in frame data
-     * 
-     * @param frameData Frame data as byte array
-     * @return List of detected tags
-     */
-    private List<AprilTagDetection> detectTags(byte[] frameData) {
-        List<AprilTagDetection> detections = new ArrayList<>();
-        
-        // In a real implementation, this would run the actual detection algorithm
-        // For now, we simulate detection with random tags
-        
-        // Simulate detection with 30% probability
-        if (random.nextDouble() < 0.3) {
-            // Generate random tag detection
-            int tagId = random.nextInt(12) + 1; // Tags 1-12
-            
-            // Random position (within reasonable range)
-            double x = (random.nextDouble() - 0.5) * 20; // -10 to 10 inches
-            double y = random.nextDouble() * 15; // 0 to 15 inches
-            double z = random.nextDouble() * 10 + 5; // 5 to 15 inches
-            
-            // Random orientation
-            double pitch = (random.nextDouble() - 0.5) * 60; // -30 to 30 degrees
-            double roll = (random.nextDouble() - 0.5) * 60;
-            double yaw = (random.nextDouble() - 0.5) * 180; // -90 to 90 degrees
-            
-            // Calculate distance and angles
-            double range = Math.sqrt(x * x + y * y + z * z);
-            double bearing = Math.atan2(x, y) * 180 / Math.PI;
-            double elevation = Math.atan2(z, Math.sqrt(x * x + y * y)) * 180 / Math.PI;
-            
-            // Create detection
-            AprilTagDetection detection = new AprilTagDetection(
-                tagId, x, y, z,
-                pitch, roll, yaw,
-                range, bearing, elevation
-            );
-            
-            detection.setConfidence(0.8 + random.nextDouble() * 0.2); // 0.8 to 1.0
-            detection.setFrameNumber((int) framesProcessed);
-            
-            detections.add(detection);
-        }
-        
-        return detections;
-    }
-    
-    // ============================================================================
-    // CONFIGURATION METHODS
-    // ============================================================================
-    
-    /**
-     * Set decimation factor
-     * 
-     * @param decimation Decimation factor (1-4)
-     */
-    public void setDecimation(int decimation) {
-        this.decimation = Math.max(1, Math.min(4, decimation));
-        
-        if (DEBUG) {
-            System.out.println("[AprilTagProcessor] Decimation set to " + this.decimation);
-        }
-    }
-    
-    /**
-     * Get current decimation factor
-     * 
-     * @return Decimation factor
-     */
-    public int getDecimation() {
-        return decimation;
-    }
-    
-    // ============================================================================
-    // STATUS METHODS
-    // ============================================================================
-    
-    /**
-     * Check if processor is initialized
-     * 
-     * @return true if initialized
-     */
-    public boolean isInitialized() {
-        return isInitialized;
-    }
-    
-    /**
-     * Get number of frames processed
-     * 
-     * @return Number of frames
-     */
-    public long getFramesProcessed() {
-        return framesProcessed;
-    }
-    
-    /**
-     * Get total tags detected
-     * 
-     * @return Number of tags detected
-     */
-    public long getTagsDetected() {
-        return tagsDetected;
-    }
-    
-    /**
-     * Get last frame processing time
-     * 
-     * @return Processing time in milliseconds
-     */
-    public long getLastProcessTime() {
-        return lastProcessTime;
-    }
-    
-    /**
-     * Get average processing time
-     * 
-     * @return Average time in milliseconds
-     */
-    public double getAverageProcessTime() {
-        if (framesProcessed == 0) return 0;
-        return lastProcessTime; // Simplified - would calculate actual average
-    }
-    
-    /**
-     * Get processor statistics
-     * 
-     * @return Statistics string
-     */
-    public String getStatistics() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("AprilTag Processor Statistics:\n");
-        sb.append("  Tag Family: ").append(TAG_FAMILY).append("\n");
-        sb.append("  Decimation: ").append(decimation).append("\n");
-        sb.append("  Frames Processed: ").append(framesProcessed).append("\n");
-        sb.append("  Tags Detected: ").append(tagsDetected).append("\n");
-        sb.append("  Last Process Time: ").append(lastProcessTime).append(" ms\n");
-        if (framesProcessed > 0) {
-            sb.append("  Average Tags/Frame: ").append(String.format("%.2f", (double) tagsDetected / framesProcessed)).append("\n");
-        }
-        return sb.toString();
-    }
-    
-    // ============================================================================
-    // CLEANUP METHODS
-    // ============================================================================
-    
-    /**
-     * Close the processor and release resources
-     */
-    public void close() {
-        if (DEBUG) {
-            System.out.println("[AprilTagProcessor] Closing processor...");
-        }
-        
-        isInitialized = false;
-        detector = null;
-        
-        if (DEBUG) {
-            System.out.println("[AprilTagProcessor] Processor closed");
-            System.out.println(getStatistics());
-        }
-    }
-}
