--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/webcam_test/src/DisplayManager.java
@@ -1,464 +1,0 @@
-/**
- * DisplayManager.java
- * 
- * Manages visualization of camera feed and AprilTag detections.
- * Handles window creation, frame rendering, and user input.
- * 
- * This class provides a visual interface for viewing the camera feed
- * and seeing detected AprilTags with their information overlaid.
- */
-
-import java.util.List;
-
-/**
- * Manages display window and visualization of detections
- */
-public class DisplayManager {
-    
-    // ============================================================================
-    // CONFIGURATION CONSTANTS
-    // ============================================================================
-    
-    /** Display window width in pixels */
-    private int windowWidth;
-    
-    /** Display window height in pixels */
-    private int windowHeight;
-    
-    /** Color for tag outlines (RGB) */
-    private static final int TAG_COLOR = 0xFF0000; // Red
-    
-    /** Color for text labels (RGB) */
-    private static final int TEXT_COLOR = 0xFFFF00; // Yellow
-    
-    /** Color for crosshair (RGB) */
-    private static final int CROSSHAIR_COLOR = 0x00FF00; // Green
-    
-    /** Font size for labels */
-    private static final int FONT_SIZE = 12;
-    
-    /** Line width for drawing */
-    private static final int LINE_WIDTH = 2;
-    
-    /** Show coordinate axes */
-    private static final boolean SHOW_AXES = true;
-    
-    /** Show crosshair at tag center */
-    private static final boolean SHOW_CROSSHAIR = true;
-    
-    /** Show tag ID */
-    private static final boolean SHOW_ID = true;
-    
-    /** Show distance information */
-    private static final boolean SHOW_DISTANCE = true;
-    
-    /** Enable debug output */
-    private static final boolean DEBUG = true;
-    
-    // ============================================================================
-    // INSTANCE VARIABLES
-    // ============================================================================
-    
-    /** Display window object (would be actual window in real implementation) */
-    private Object displayWindow;
-    
-    /** Flag indicating if display is initialized */
-    private boolean isInitialized;
-    
-    /** Flag to signal quit */
-    private boolean shouldQuit;
-    
-    /** Frame counter for display */
-    private long frameCount;
-    
-    /** Last frame display time */
-    private long lastFrameTime;
-    
-    // ============================================================================
-    // CONSTRUCTOR
-    // ============================================================================
-    
-    /**
-     * Create a new DisplayManager
-     * 
-     * @param width Display window width
-     * @param height Display window height
-     */
-    public DisplayManager(int width, int height) {
-        this.windowWidth = width;
-        this.windowHeight = height;
-        this.isInitialized = false;
-        this.shouldQuit = false;
-        this.frameCount = 0;
-        this.lastFrameTime = System.currentTimeMillis();
-        
-        if (DEBUG) {
-            System.out.println("[DisplayManager] Created with resolution " + width + "x" + height);
-        }
-    }
-    
-    // ============================================================================
-    // INITIALIZATION METHODS
-    // ============================================================================
-    
-    /**
-     * Initialize the display window
-     * 
-     * @return true if initialization successful, false otherwise
-     */
-    public boolean initialize() {
-        try {
-            if (DEBUG) {
-                System.out.println("[DisplayManager] Initializing display window...");
-            }
-            
-            // In a real implementation, this would:
-            // 1. Create a window using Swing or JavaFX
-            // 2. Set window size and title
-            // 3. Create graphics context
-            // 4. Set up event listeners
-            
-            if (!createWindow()) {
-                System.err.println("[DisplayManager] Failed to create window");
-                return false;
-            }
-            
-            isInitialized = true;
-            
-            if (DEBUG) {
-                System.out.println("[DisplayManager] Display window initialized successfully");
-                System.out.println("  - Window Size: " + windowWidth + "x" + windowHeight);
-                System.out.println("  - Press 'Q' to quit");
-            }
-            
-            return true;
-            
-        } catch (Exception e) {
-            System.err.println("[DisplayManager] Initialization error: " + e.getMessage());
-            e.printStackTrace();
-            return false;
-        }
-    }
-    
-    /**
-     * Create the display window
-     * 
-     * @return true if successful
-     */
-    private boolean createWindow() {
-        try {
-            // In a real implementation, this would create an actual window
-            // For now, we simulate it
-            displayWindow = new Object(); // Placeholder
-            
-            if (DEBUG) {
-                System.out.println("[DisplayManager] Display window created");
-            }
-            
-            return true;
-            
-        } catch (Exception e) {
-            System.err.println("[DisplayManager] Window creation failed: " + e.getMessage());
-            return false;
-        }
-    }
-    
-    // ============================================================================
-    // DISPLAY METHODS
-    // ============================================================================
-    
-    /**
-     * Display a frame with detected AprilTags
-     * 
-     * @param frameData Frame data as byte array
-     * @param detections List of detected AprilTags
-     */
-    public void displayFrame(byte[] frameData, List<AprilTagDetection> detections) {
-        if (!isInitialized || frameData == null) {
-            return;
-        }
-        
-        try {
-            // In a real implementation, this would:
-            // 1. Convert frame data to displayable format
-            // 2. Draw frame on window
-            // 3. Draw detection overlays
-            // 4. Update window
-            
-            drawFrame(frameData);
-            drawDetections(detections);
-            updateDisplay();
-            
-            frameCount++;
-            lastFrameTime = System.currentTimeMillis();
-            
-        } catch (Exception e) {
-            System.err.println("[DisplayManager] Display error: " + e.getMessage());
-        }
-    }
-    
-    /**
-     * Draw the camera frame
-     * 
-     * @param frameData Frame data as byte array
-     */
-    private void drawFrame(byte[] frameData) {
-        // In a real implementation, this would:
-        // 1. Convert frame data to image
-        // 2. Draw image on graphics context
-        // 3. Handle color space conversion if needed
-        
-        if (DEBUG && frameCount % 30 == 0) {
-            System.out.println("[DisplayManager] Drawing frame " + frameCount);
-        }
-    }
-    
-    /**
-     * Draw detection overlays
-     * 
-     * @param detections List of detected AprilTags
-     */
-    private void drawDetections(List<AprilTagDetection> detections) {
-        if (detections == null || detections.isEmpty()) {
-            return;
-        }
-        
-        // In a real implementation, this would:
-        // 1. For each detection:
-        //    a. Draw tag outline
-        //    b. Draw tag ID
-        //    c. Draw distance information
-        //    d. Draw coordinate axes
-        //    e. Draw crosshair
-        
-        for (AprilTagDetection detection : detections) {
-            drawTagOutline(detection);
-            drawTagInfo(detection);
-            
-            if (SHOW_AXES) {
-                drawCoordinateAxes(detection);
-            }
-            
-            if (SHOW_CROSSHAIR) {
-                drawCrosshair(detection);
-            }
-        }
-    }
-    
-    /**
-     * Draw tag outline
-     * 
-     * @param detection Detection to draw
-     */
-    private void drawTagOutline(AprilTagDetection detection) {
-        // In a real implementation, this would:
-        // 1. Calculate tag corners in image space
-        // 2. Draw rectangle outline
-        // 3. Use TAG_COLOR and LINE_WIDTH
-        
-        if (DEBUG && frameCount % 100 == 0) {
-            System.out.println("[DisplayManager] Drawing tag outline for ID " + detection.getId());
-        }
-    }
-    
-    /**
-     * Draw tag information
-     * 
-     * @param detection Detection to draw
-     */
-    private void drawTagInfo(AprilTagDetection detection) {
-        // In a real implementation, this would:
-        // 1. Format information string
-        // 2. Draw text on image
-        // 3. Use TEXT_COLOR and FONT_SIZE
-        
-        StringBuilder info = new StringBuilder();
-        
-        if (SHOW_ID) {
-            info.append("ID: ").append(detection.getId());
-        }
-        
-        if (SHOW_DISTANCE) {
-            if (info.length() > 0) info.append(" | ");
-            info.append(String.format("Dist: %.1f\"", detection.getRange()));
-        }
-        
-        if (DEBUG && frameCount % 100 == 0) {
-            System.out.println("[DisplayManager] Tag info: " + info.toString());
-        }
-    }
-    
-    /**
-     * Draw coordinate axes
-     * 
-     * @param detection Detection to draw
-     */
-    private void drawCoordinateAxes(AprilTagDetection detection) {
-        // In a real implementation, this would:
-        // 1. Calculate axis endpoints in image space
-        // 2. Draw X, Y, Z axes with different colors
-        // 3. Use appropriate colors and line width
-        
-        if (DEBUG && frameCount % 100 == 0) {
-            System.out.println("[DisplayManager] Drawing coordinate axes");
-        }
-    }
-    
-    /**
-     * Draw crosshair at tag center
-     * 
-     * @param detection Detection to draw
-     */
-    private void drawCrosshair(AprilTagDetection detection) {
-        // In a real implementation, this would:
-        // 1. Calculate tag center in image space
-        // 2. Draw crosshair lines
-        // 3. Use CROSSHAIR_COLOR
-        
-        if (DEBUG && frameCount % 100 == 0) {
-            System.out.println("[DisplayManager] Drawing crosshair");
-        }
-    }
-    
-    /**
-     * Update the display window
-     */
-    private void updateDisplay() {
-        // In a real implementation, this would:
-        // 1. Refresh the window
-        // 2. Render all drawn elements
-        // 3. Handle window events
-    }
-    
-    // ============================================================================
-    // INPUT HANDLING METHODS
-    // ============================================================================
-    
-    /**
-     * Check if user requested quit
-     * 
-     * @return true if quit requested
-     */
-    public boolean shouldQuit() {
-        // In a real implementation, this would:
-        // 1. Check for keyboard input
-        // 2. Check for window close button
-        // 3. Return true if quit requested
-        
-        // Simulate keyboard input checking
-        checkKeyboardInput();
-        
-        return shouldQuit;
-    }
-    
-    /**
-     * Check for keyboard input
-     */
-    private void checkKeyboardInput() {
-        // In a real implementation, this would:
-        // 1. Check for 'Q' or 'q' key press
-        // 2. Check for ESC key press
-        // 3. Set shouldQuit flag if detected
-        
-        // For now, we don't actually check input
-        // In a real implementation, this would use KeyListener or similar
-    }
-    
-    /**
-     * Request quit
-     */
-    public void requestQuit() {
-        shouldQuit = true;
-        if (DEBUG) {
-            System.out.println("[DisplayManager] Quit requested");
-        }
-    }
-    
-    // ============================================================================
-    // CONFIGURATION METHODS
-    // ============================================================================
-    
-    /**
-     * Set window size
-     * 
-     * @param width Window width
-     * @param height Window height
-     */
-    public void setWindowSize(int width, int height) {
-        this.windowWidth = width;
-        this.windowHeight = height;
-        
-        if (DEBUG) {
-            System.out.println("[DisplayManager] Window size set to " + width + "x" + height);
-        }
-    }
-    
-    /**
-     * Set window title
-     * 
-     * @param title Window title
-     */
-    public void setWindowTitle(String title) {
-        // In a real implementation, this would set the window title
-        
-        if (DEBUG) {
-            System.out.println("[DisplayManager] Window title set to: " + title);
-        }
-    }
-    
-    // ============================================================================
-    // STATUS METHODS
-    // ============================================================================
-    
-    /**
-     * Check if display is initialized
-     * 
-     * @return true if initialized
-     */
-    public boolean isInitialized() {
-        return isInitialized;
-    }
-    
-    /**
-     * Get number of frames displayed
-     * 
-     * @return Number of frames
-     */
-    public long getFrameCount() {
-        return frameCount;
-    }
-    
-    /**
-     * Get display statistics
-     * 
-     * @return Statistics string
-     */
-    public String getStatistics() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("Display Statistics:\n");
-        sb.append("  Window Size: ").append(windowWidth).append("x").append(windowHeight).append("\n");
-        sb.append("  Frames Displayed: ").append(frameCount).append("\n");
-        return sb.toString();
-    }
-    
-    // ============================================================================
-    // CLEANUP METHODS
-    // ============================================================================
-    
-    /**
-     * Close the display window and release resources
-     */
-    public void close() {
-        if (DEBUG) {
-            System.out.println("[DisplayManager] Closing display window...");
-        }
-        
-        isInitialized = false;
-        displayWindow = null;
-        
-        if (DEBUG) {
-            System.out.println("[DisplayManager] Display window closed");
-            System.out.println(getStatistics());
-        }
-    }
-}
